refactor: 변수명, 컴포넌트명 변경
