+++ src/components/DataTable/DataTable.module.scss
@@ -19,7 +19,7 @@ $normal: #666666;
     font-weight: 700;
   }
 
-  &__sub-title {
+  &__subtitle {
     color: $normal;
     font-size: 16px;
     font-weight: 400;
