+++ src/components/editor/TextEditor/TextEditor.module.scss
@@ -103,7 +103,7 @@
     margin-bottom: 4px;
   }
 
-  &__sub-title {
+  &__subtitle {
     color: #898a8d;
     font-weight: 400;
     font-size: 14px;
