+++ src/pages/Home/Home.module.scss
@@ -1,4 +1,4 @@
-.locationBox {
+.location-box {
   position: absolute;
   top: 0;
   left: 0;
