+++ src/pages/Home/components/Map/components/LocationSelectModal/LocationSelectModal.module.scss
@@ -26,7 +26,7 @@
     margin: 68px auto 12px;
   }
 
-  &__subTitle {
+  &__subtitle {
     display: flex;
     font-size: 16px;
     font-weight: 400;
