+++ src/pages/Home/index.tsx
@@ -1,5 +1,5 @@
-import LocationInfo from 'pages/Home/components/LocationInfo/index';
-import Location from 'pages/Home/components/Map/components/Location/index';
+import LocationSelectButton from 'pages/Home/components/LocationSelectButton/index';
+import LocationSelectModal from 'pages/Home/components/Map/components/LocationSelectModal/index';
 import useHome from 'pages/Home/components/Map/hooks/useHome';
 import Map from 'pages/Home/components/Map/index';
 
@@ -7,18 +7,19 @@ import styles from './Home.module.scss';
 
 export default function Home(): JSX.Element {
   const {
-    isClickLocation, active, userLocation, locationRef,
+    isClickLocation, activate, userLocation, locationRef,
   } = useHome();
+
   return (
     <>
-      <LocationInfo address={userLocation.address} onClick={active} />
       <Map />
+      <LocationSelectButton address={userLocation.address} onClick={activate} />
       {isClickLocation && (
         <div
-          className={styles.locationBox}
+          className={styles['location-box']}
           ref={locationRef}
         >
-          <Location />
+          <LocationSelectModal />
         </div>
       )}
     </>
