+++ src/pages/Notice/Notice.module.scss
@@ -38,7 +38,7 @@ $normal: #666666;
     margin: 156px auto 10px;
   }
 
-  &__sub-title {
+  &__subtitle {
     color: $normal;
     font-size: 16px;
     font-weight: 400;
