+++ src/pages/Notice/index.tsx
@@ -6,7 +6,7 @@ import styles from './Notice.module.scss';
 export default function Notice(): JSX.Element {
   const { flatData: postData } = usePostList({ size: 10 });
   const title = '공지사항';
-  const subTitle = '쩝쩝박사가 여러분들께 전달드립니다!';
+  const subtitle = '쩝쩝박사가 여러분들께 전달드립니다!';
 
   return (
     <div className={styles.container}>
@@ -16,7 +16,7 @@ export default function Notice(): JSX.Element {
             <Datatable
               data={postData.content}
               title={title}
-              subTitle={subTitle}
+              subtitle={subtitle}
             />
           )}
       </div>
