+++ src/pages/Setting/Mobile/Setting.module.scss
@@ -36,7 +36,7 @@ $back: white;
   flex-direction: column;
   font-weight: 500;
 
-  &__sub-title {
+  &__subtitle {
     display: flex;
     font-size: 14px;
     margin: 0 290px 16.5px 17px;
@@ -77,7 +77,7 @@ $back: white;
   margin-top: 36px;
   margin-right: 15px;
 
-  &__sub-title {
+  &__subtitle {
     display: flex;
     font-size: 14px;
     margin: 0 290px 16.5px 17px;
