+++ src/pages/Setting/Mobile/index.tsx
@@ -23,7 +23,7 @@ export default function MobileSetting() {
         <div className={styles.header__title}>설정</div>
       </div>
       <div className={styles.account}>
-        <div className={styles['account__sub-title']}>계정 관리</div>
+        <div className={styles.account__subtitle}>계정 관리</div>
         <div className={styles.account__content}>
           <div className={styles.account__text}>비밀번호 변경 </div>
           <Link to="id-Change">
@@ -42,7 +42,7 @@ export default function MobileSetting() {
         </Link>
       </div>
       <div className={styles.service}>
-        <div className={styles['service__sub-title']}>서비스</div>
+        <div className={styles.service__subtitle}>서비스</div>
         <div className={styles.service__content}>
           <div className={styles.service__text}>공지사항</div>
           <Link to="/notice">
